pyproject.toml
[tool.poetry]
name = "chitr"
version = "0.1.0"
description = ""
authors = ["mohitkr05 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
